+++ src/main.rs
@@ -29,16 +29,10 @@ mod log_command;
 mod revision;
 mod repository;
 
-use flate2::Compression;
 use flate2::write::ZlibDecoder;
-use flate2::write::ZlibEncoder;
-use hex_string::HexString;
-use sha1::Digest;
-use sha1::Sha1;
 use std::env;
 use std::fs;
 use std::io::Write;
-use std::path::Path;
 use std::path::PathBuf;
 
 use crate::command::Command;
@@ -343,70 +337,4 @@ fn process_garbage_collection() -> Result<(), ZatsuError> {
     }
 
     Ok(())
-}
-
-fn process_file(path: impl AsRef<Path>) -> Result<String, ZatsuError> {
-    let metadata = match fs::metadata(&path) {
-        Ok(metadata) => metadata,
-        Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_READING_META_DATA_FAILED)),
-    };
-    let mut hex_string = String::new();
-    if metadata.is_file() {
-        println!("This is file.");
-        let values = match fs::read(path) {
-            Ok(values) => values,
-            Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_LOADING_FILE_FAILED)),
-        };
-        println!("{} bytes read.", values.len());
-        let mut sha1 = Sha1::new();
-        sha1.update(values.clone());
-        let hash = sha1.finalize();
-        let hash_values = hash.to_vec();
-        println!("{} bytes of hash generated.", hash_values.len());
-        // println!("{}", hash_values);
-        let hex = HexString::from_bytes(&hash_values);
-        hex_string = hex.as_string();
-        println!("{}", hex_string);
-
-        let directory_name = hex_string[0..2].to_string();
-        let path = format!(".zatsu/objects/{}", directory_name).to_string();
-        let a_path = Path::new(&path);
-        let exists = match a_path.try_exists() {
-            Ok(exists) => exists,
-            Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_SAVING_FILE_FAILED)),
-        };
-        if !exists {
-            match fs::create_dir(&path) {
-                Ok(()) => (),
-                Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_SAVING_FILE_FAILED)),
-            };
-        }
-
-        let path = format!("{}/{}", &path, hex_string);
-        let a_path = Path::new(&path);
-        let exists = match a_path.try_exists() {
-            Ok(exists) => exists,
-            Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_SAVING_FILE_FAILED)),
-        };
-        if !exists {
-            let mut encoder = ZlibEncoder::new(Vec::new(), Compression::default());
-            match encoder.write_all(&values) {
-                Ok(()) => (),
-                Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_SAVING_FILE_FAILED)),
-            }
-            let compressed = match encoder.finish() {
-                Ok(compressed) => compressed,
-                Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_SAVING_FILE_FAILED)),
-            };
-
-            match fs::write(path, compressed) {
-                Ok(()) => (),
-                Err(_) => return Err(ZatsuError::new("main".to_string(), ERROR_SAVING_FILE_FAILED)),
-            };
-        }
-    } else {
-        println!("This is not file.");
-    }
-
-    Ok(hex_string)
 }
